# docs/swagger.yaml
basePath: /api/v1
definitions:
  ginx.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  ginx.ResponseFail:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
  schema.GetProfileResp:
    properties:
      posts:
        items:
          $ref: '#/definitions/schema.SysPost'
        type: array
      roles:
        items:
          $ref: '#/definitions/schema.SysRole'
        type: array
      user:
        $ref: '#/definitions/schema.SysUser'
    type: object
...
  schema.PassWord:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  schema.ResetSysUserPwdReq:
    properties:
      password:
        type: string
      updateBy:
        type: integer
      userId:
        type: integer
    type: object
...
  schema.SysApi:
    properties:
      action:
        description: 请求类型
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      deletedAt:
        description: 删除时间
        type: string
      handle:
        description: handle
        type: string
      id:
        description: 主键编码
        type: integer
      path:
        description: 地址
        type: string
      title:
        description: 标题
        type: string
      type:
        description: 接口类型
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
...
  schema.SysDept:
    properties:
      children:
        items:
          $ref: '#/definitions/schema.SysDept'
        type: array
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dataScope:
        type: string
      deptId:
        description: 部门编码
        type: integer
      deptName:
        description: 部门名称
        type: string
      deptPath:
        type: string
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        type: string
      params:
        type: string
      parentId:
        description: 上级部门
        type: integer
      phone:
        description: 手机
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
...
  schema.SysMenu:
    properties:
      action:
        type: string
      apis:
        items:
          type: integer
        type: array
      breadcrumb:
        type: integer
      children:
        items:
          $ref: '#/definitions/schema.SysMenu'
        type: array
      component:
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dataScope:
        type: string
      icon:
        type: string
      is_select:
        type: boolean
      isFrame:
        type: integer
      menuId:
        type: integer
      menuName:
        type: string
      menuType:
        type: string
      noCache:
        type: integer
      params:
        type: string
      parentId:
        type: integer
      path:
        type: string
      paths:
        type: string
      permission:
        type: string
      roleId:
        type: integer
      sort:
        type: integer
      sysApi:
        items:
          $ref: '#/definitions/schema.SysApi'
        type: array
      title:
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
      visible:
        type: integer
    type: object
...
  schema.SysPost:
    properties:
      createBy:
        description: 创建者
        type: integer
      dataScope:
        type: string
      params:
        type: string
      postCode:
        description: 岗位代码
        type: string
      postId:
        description: 岗位编号
        type: integer
      postName:
        description: 岗位名称
        type: string
      remark:
        description: 描述
        type: string
      sort:
        description: 岗位排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
    type: object
...
  schema.SysRole:
    properties:
      admin:
        type: integer
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dataScope:
        type: string
      flag:
        type: string
      menuIds:
        items:
          type: integer
        type: array
      params:
        type: string
      remark:
        description: 备注
        type: string
      roleId:
        description: 角色编码
        type: integer
      roleKey:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 角色排序
        type: integer
      status:
        type: integer
      sysMenu:
        items:
          $ref: '#/definitions/schema.SysMenu'
        type: array
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
...
  schema.SysUser:
    properties:
      avatar:
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dept:
        $ref: '#/definitions/schema.SysDept'
      deptId:
        type: integer
      deptIds:
        items:
          type: integer
        type: array
      email:
        type: string
      nickName:
        type: string
      phone:
        type: string
      postId:
        type: integer
      postIds:
        items:
          type: integer
        type: array
      remark:
        type: string
      roleId:
        type: integer
      roleIds:
        items:
          type: integer
        type: array
      sex:
        type: integer
      status:
        type: integer
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
  schema.SysUserDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  schema.SysUserGetPageResp:
    properties:
      data:
        items:
          $ref: '#/definitions/schema.SysUser'
        type: array
      total:
        type: integer
    type: object
  schema.SysUserInsertReq:
    properties:
      avatar:
        type: string
      createBy:
        type: integer
      deptId:
        type: integer
      email:
        type: string
      nickName:
        type: string
      password:
        type: string
      phone:
        type: string
      postId:
        type: integer
      remark:
        type: string
      roleId:
        type: integer
      sex:
        type: integer
      status:
        default: 2
        type: integer
      username:
        type: string
    required:
    - deptId
    - email
    - nickName
    - password
    - phone
    - status
    - username
    type: object
  schema.SysUserUpdateReq:
    properties:
      avatar:
        type: string
      deptId:
        type: integer
      email:
        type: string
      nickName:
        type: string
      phone:
        type: string
      postId:
        type: integer
      remark:
        type: string
      roleId:
        type: integer
      sex:
        type: integer
      status:
        type: integer
      updateBy:
        type: integer
      userId:
        type: integer
      username:
        type: string
    required:
    - deptId
    - email
    - nickName
    - phone
    - username
    type: object
...
  schema.UpdateSysUserStatusReq:
    properties:
      status:
        type: integer
      updateBy:
        type: integer
      userId:
        type: integer
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: conjurer888888@gmail.com
    name: conjurer
    url: https:/github.com/dot123
  description: This is a game management background. you can use the api key `ApiKeyAuth`
    to test the authorization filters.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com
  title: gin-gorm-admin API
  version: "1.0"
paths:
...
  /sys-user:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysUserDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 删除用户数据
      tags:
      - 用户
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: createdAtOrder
        type: string
      - in: query
        name: deptId
        type: integer
      - in: query
        name: email
        type: string
      - in: query
        name: nickName
        type: string
      - in: query
        name: pageNum
        type: integer
      - in: query
        maximum: 100
        name: pageSize
        type: integer
      - in: query
        name: phone
        type: string
      - in: query
        name: postID
        type: integer
      - in: query
        name: roleID
        type: integer
      - in: query
        name: sex
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: statusOrder
        type: string
      - in: query
        name: userID
        type: integer
      - in: query
        name: userIdOrder
        type: string
      - in: query
        name: username
        type: string
      - in: query
        name: usernameOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysUserGetPageResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 列表用户信息数据
      tags:
      - 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysUserInsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 创建用户
      tags:
      - 用户
    put:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysUserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改用户数据
      tags:
      - 用户
  /sys-user/{userId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysUser'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取用户
      tags:
      - 用户
  /user/avatar:
    post:
      consumes:
      - application/json
      - multipart/form-data
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改头像
      tags:
      - 个人中心
  /user/getinfo:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取个人信息
      tags:
      - 个人中心
  /user/profile/{userId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: userId
        in: path
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetProfileResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取个人中心用户
      tags:
      - 个人中心
  /user/pwd/reset:
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ResetSysUserPwdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 重置用户密码
      tags:
      - 用户
  /user/pwd/set:
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.PassWord'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 用户
  /user/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateSysUserStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改用户状态
      tags:
      - 个人中心
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"